==> src/s04_correlateKmerAndSNPvariation.Rmd <==
---
title: "K-mer approach to Lewontin's paradox results"
author: "Miles Roberts"
date: "2024-09-08"
output:
  html_document:
    df_print: paged
  pdf_document: default
---

# load packages
```{r}
rm(list = ls())

# remove.packages(c("dplyr", "ggplot2", "caper", "stringr", "data.table", "cowplot", "rWCVP", "lwgeom", "sf"))

library(dplyr)
library(ggplot2)
library(caper)
library(stringr)
library(data.table)
#library(cowplot)
#library(scico)
#library(giscoR)
library(rWCVP)
library(lwgeom)
library(sf)
library(ggpubr)
library(broom)


today = Sys.Date()
```

# Data preparation

## Prepare diversity data 
```{r}
# List input files
pi = list.files("../results/2023-03-26/genome-wide-pi", full.names = T)
bcd = list.files("../results/2023-03-26/bray-curtis-dissimilarities", full.names = T)
jac = list.files("../results/2023-03-26/jaccard-dissimilarities", full.names = T)

# for now, remove species that can't be processed further for whatever reason
# pi: panicum virgatum
# jac and bcd: panicum hallii, lens culinaris
jac = jac[c(-56, -81)]
bcd = bcd[c(-56, -81)]
pi = pi[c(-80)]

# get species IDs
species = gsub("_pairwise-pi.txt", "", gsub("../results/2023-03-26/genome-wide-pi/", "", pi))

# Load input files
pi = lapply(pi, fread, header = T, data.table = F)
bcd = lapply(bcd, fread, header = F, data.table = F)
jac = lapply(jac, fread, header = F, data.table = F)

# calculate mean pairwise dissimilarity
bcdMean = lapply(bcd, function(x){mean(x[,2])})
jacMean = lapply(jac, function(x){mean(x[,2])})

# calculate average number of k-mer differences as union - intersection
unionMean = lapply(jac, function(x){sum(x[,4]-x[,3])})

# calculate number of individuals sampled for each dataset
# You get this equation by rewriting the formula for the number of pairwise comparisons (n^2 - n)/2 = N as a polynomial and applying the quadratic formula
backCalcN = function(x){
  sqrt(1/4 + 2*nrow(x)) + 1/2
}

nidv = lapply(bcd, backCalcN)

# cbind all diversity data into one dataframe per species
alldiv = Map(cbind, pi, bcdMean, jacMean, unionMean, nidv)

# Some species have extra data, remove for now
speciesWithExtraData = which(unlist(lapply(alldiv, ncol)) == 12)
for(i in speciesWithExtraData){
  alldiv[[i]] = alldiv[[i]][,c(1, 3:6, 9:12)]
}

# rbind all species data into one dataframe
alldiv = lapply(alldiv, setNames, c("h", "ntotal", "nvariant", "ninvariant", "pi", "bcd", "jac", "umean", "nidv"))

alldiv = do.call(rbind, alldiv)

# add species names
alldiv$species = species
alldiv$species = str_to_sentence(gsub("_", " ", alldiv$species))

# For now, remove species where invariant sites were not properly called
# alldiv = alldiv[(alldiv$ninvariant > 0),]

# For now, remove species where variants sites were not properly called
# alldiv = alldiv[(alldiv$nvariant > 30),]

# remove species where there aren't many sampled individuals
# alldiv = alldiv[(alldiv$nidv >= 10),]

# add coverage covariate
coverage = read.table("../results/2023-03-26/bp_sequenced_2024-04-03.txt", header = T, sep = "\t")
alldiv = merge(alldiv, coverage, by = "species", all.x = TRUE)

# remove temporary objects
rm(pi, jac, bcd, coverage, bcdMean, jacMean, unionMean, nidv)

# 
# gsize = read.table("../results/2023-03-26/genome_size.txt", header = T, sep = "\t")
# alldiv = merge(alldiv, gsize, by = "species")
# 
# alldiv$totalcov = alldiv$totalbp/alldiv$assemblySize
```

## Load range data
```{r}
# load range data
areas = list.files("../results/2023-03-26/ranges/areas", full.names = T)

# remove files with no data
# Cicer arietinum
#areas = areas[c(-33)]

# extract species names
areas_species = gsub("_areas.txt", "", gsub("../results/2023-03-26/ranges/areas/", "", areas))
areas_species = str_to_sentence(gsub("_", " ", areas_species))

# load data
areas = lapply(areas, read.table, header = +T, sep = "\t")

# add names to list
names(areas) = areas_species

# copy object so that we can separate out whether invaded areas are included or not
native_areas = areas
```

## remove areas outside of native range 
```{r}
# Refer to plants of the world online for native range maps
# powo.science.kew.org

# Amaranthus hybridus native to only North and South America
native_areas$`Amaranthus hypochondriacus` = native_areas$`Amaranthus hypochondriacus`[1:2,]

# Arabidopsis thaliana is only native to Europe, Asia, and Africa
native_areas$`Arabidopsis thaliana` = native_areas$`Arabidopsis thaliana`[2:4,]

# Arachis duranensis is native to south america, remove asia
native_areas$`Arachis duranensis` = native_areas$`Arachis duranensis`[1,]

# Arachis glabrata is native to Brazil, remove North America
native_areas$`Arachis hypogaea` = native_areas$`Arachis hypogaea`[2,]

# Beta vulgaris subsp. vulgaris is native to only Europe and Africa
native_areas$`Beta vulgaris` = native_areas$`Beta vulgaris`[2:3,]

# Brachypodium distachyon is native to Europe, Asia, Africa
native_areas$`Brachypodium distachyon` = native_areas$`Brachypodium distachyon`[2:4,]

# Brassica juncea is native only to Europe
native_areas$`Brassica juncea` = native_areas$`Brassica juncea`[4,]

# Brassica cretica native to europe
native_areas$`Brassica oleracea` = native_areas$`Brassica oleracea`[1,]

# Brassica fruticulosa native to europe
native_areas$`Brassica rapa` = native_areas$`Brassica rapa`[2,]

# Buddleja alternifolia is native to Asia
native_areas$`Buddleja alternifolia` = native_areas$`Buddleja alternifolia`[2,]

# Cajanus scarabaeoides native to Africa, Asia, Australia
native_areas$`Cajanus cajan` = native_areas$`Cajanus cajan`[c(1,3,4),]

# Camelina sativa is not native to North America
native_areas$`Camelina sativa` = native_areas$`Camelina sativa`[c(2,3),]

# Cannabis sativa is native only to Asia
native_areas$`Cannabis sativa` = native_areas$`Cannabis sativa`[5,]

# Capsicum annuum is native to north america and south america
native_areas$`Capsicum annuum` = native_areas$`Capsicum annuum`[c(1,2),]

# Castanea mollissima is native only to asia
native_areas$`Castanea mollissima` = native_areas$`Castanea mollissima`[2,]

# Corylus americana is only native to north america
native_areas$`Corylus americana` = native_areas$`Corylus americana`[1,]

# Cucumis myriocarpus is native only to africa
native_areas$`Cucumis melo` = native_areas$`Cucumis melo`[2,]

# Cucurbita palmata is only native to North America
native_areas$`Cucurbita pepo` = native_areas$`Cucurbita pepo`[1,]

# Digitaria longiflora native to africa, asia, australia
native_areas$`Digitaria exilis`= native_areas$`Digitaria exilis`[2:4,]

# Dioscorea communis is native only to africa
native_areas$`Dioscorea rotundata` = native_areas$`Dioscorea rotundata`[2,]

# Gossypium arboreum native to asia
native_areas$`Gossypium arboreum` = native_areas$`Gossypium arboreum`[3,]

# Gossypium thurberi is native only to north america
native_areas$`Gossypium hirsutum` = native_areas$`Gossypium hirsutum`[1,]

# Helianthus maximiliani is native only to north america
native_areas$`Helianthus annuus` = native_areas$`Helianthus annuus`[1,]

# Juglans mandshuica is native only to asia
native_areas$`Juglans regia` = native_areas$`Juglans regia`[2,]

# Lactuca sativa is native to Iraq
native_areas$`Lactuca sativa` = native_areas$`Lactuca sativa`[3,]

# Linum bienne is native to europe, africa, and asia
native_areas$`Linum usitatissimum` = native_areas$`Linum usitatissimum`[c(3,4),]

# Malus sylvestris is only native to Europe
native_areas$`Malus sylvestris` = native_areas$`Malus sylvestris`[2,]

# Medicago truncatula only native to europe and africa
native_areas$`Medicago truncatula` = native_areas$`Medicago truncatula`[c(4,5),]

# Mimulus guttatus only native to North America
native_areas$`Mimulus guttatus` = native_areas$`Mimulus guttatus`[1,]

# Miscanthus sacchariflorus is native only to asia
native_areas$`Miscanthus sinensis` = native_areas$`Miscanthus sinensis`[3,]

# Musa relative only native to asia and india
native_areas$`Musa acuminata` = native_areas$`Musa acuminata`[2,]

# Nicotiana glauca is native only to south america
native_areas$`Nicotiana tabacum` = native_areas$`Nicotiana tabacum`[2,]

# Oryza rufipogon is native to asia and australia
native_areas$`Oryza rufipogon` = native_areas$`Oryza rufipogon`[c(2,3),]

# Panicum hallii is native only to north america
native_areas$`Panicum hallii` = native_areas$`Panicum hallii`[1,]

# Panicum capillare is native only to north america
native_areas$`Panicum virgatum` = native_areas$`Panicum virgatum`[1,]

# Papaver rhoeas is native to Africa, Europe, Asia
native_areas$`Papaver somniferum` = native_areas$`Papaver somniferum`[4:6,]

# Populus deltoides only native to north america
native_areas$`Populus deltoides` = native_areas$`Populus deltoides`[1,]

# Prunus avium is native to europe, asia, africa
native_areas$`Prunus avium` = native_areas$`Prunus avium`[2:3,]

# Raphanus raphanistrum is native to europe and africa
native_areas$`Raphanus sativus` = native_areas$`Raphanus sativus`[3:4,]

# Setaria faberi is native only to asia
native_areas$`Setaria italica` = native_areas$`Setaria italica`[3,]

# Setaria viridis is native to europe, africa, asia, australia
native_areas$`Setaria viridis` = native_areas$`Setaria viridis`[2:5,]

# Thlaspi arvense is native to europe and asia
native_areas$`Thlaspi arvense` = native_areas$`Thlaspi arvense`[3:4,]

# Triticum urartu is native only to asia 
native_areas$`Triticum aestivum` = native_areas$`Triticum aestivum`[2,]
```

## Merge range data with diversity data
```{r}
# function to get non-water area
get_nonwater_area = function(x){
  sum(x[,"total_area"] - x[,"water_area"])
}

# format area data as frame
areas_frame = data.frame(
  species = areas_species,
  gbif_all_area = unlist(lapply(areas, get_nonwater_area)),
  gbif_native_area = unlist(lapply(native_areas, get_nonwater_area))
)

# manually correct gossypium barbadense because galapagos was incorrectly counted as water
areas_frame$gbif_all_area[(areas_frame$species == "Gossypium barbadense")] = areas$`Gossypium barbadense`$total_area
areas_frame$gbif_native_area[(areas_frame$species == "Gossypium barbadense")] = areas$`Gossypium barbadense`$total_area

# merge with diversity data
alldiv = merge(alldiv, areas_frame, by = "species", all.x = T)
print(alldiv)

# save final table
#write.csv(alldiv, paste("../results/2023-03-26/diversityAndRangeData_", today, ".csv", sep = ""), row.names = F, quote = F)
```



## Load range areas
```{r}
# load local copies of World Checklist of Vascular Plants
wcvp_dist = fread("../workflow/data/range/wcvp_distribution.csv.gz", sep = "|", header = T, quote = "", fill = T, encoding = "UTF-8")

wcvp_names = fread("../workflow/data/range/wcvp_names.csv.gz", sep = "|", header = T, quote = "", fill = T, encoding = "UTF-8")

# update species names to match latest names in WCVP
updated_species_names = data.frame(species = unique(alldiv$species),species_updated = unique(alldiv$species))

updated_species_names$species_updated[updated_species_names$species == "Arabis nemorensis"] = "Arabis planisiliqua"
updated_species_names$species_updated[updated_species_names$species == "Dioscorea rotundata"] = "Dioscorea cayenensis subsp. rotundata"
updated_species_names$species_updated[updated_species_names$species == "Glycine soja"] = "Glycine max subsp. soja"
updated_species_names$species_updated[updated_species_names$species == "Lens ervoides"] = "Vicia lenticula"
updated_species_names$species_updated[updated_species_names$species == "Mimulus guttatus"] = "Erythranthe guttata"
updated_species_names$species_updated[updated_species_names$species == "Populus trichocarpa"] = "Populus tristis"
updated_species_names$species_updated[updated_species_names$species == "Raphanus sativus"] = "Raphanus raphanistrum subsp. sativus"
updated_species_names$species_updated[updated_species_names$species == "Salix dunnii"] = "Salix tetrasperma"
updated_species_names$species_updated[updated_species_names$species == "Solanum stenotomum"] = "Solanum tuberosum"
updated_species_names$species_updated[updated_species_names$species == "Boechera stricta"] = "Boechera drummondii"

# loop over each species, and calculate range area in square kilometers
region_areas = NULL

for(species in updated_species_names$species_updated){

  print(species)
  
  # get native distribution
species_dist = wcvp_distribution(
taxon = species,
taxon_rank = "species",
native = TRUE,
introduced = FALSE,
extinct = TRUE,
location_doubtful = FALSE,
wcvp_distributions = wcvp_dist,
wcvp_names = wcvp_names
)

# get native + invaded species distribution
species_dist_invaded = wcvp_distribution(
taxon = species,
taxon_rank = "species",
native = TRUE,
introduced = TRUE,
extinct = TRUE,
location_doubtful = FALSE,
wcvp_distributions = wcvp_dist,
wcvp_names = wcvp_names
)

# calculate area, convert to square kilometers (1000 x 1000 meters)
region_area = sum(st_area(species_dist))/1e6
region_area_invaded = sum(st_area(species_dist_invaded))/1e6

region_areas = rbind(region_areas, data.frame(species_updated = species, wcvp_native_area = region_area, wcvp_all_area = region_area_invaded))

}

# merge range areas with old species names
region_areas = merge(updated_species_names, region_areas, by = "species_updated")

# remove weird units
region_areas$wcvp_native_area = as.numeric(region_areas$wcvp_native_area)
region_areas$wcvp_all_area = as.numeric(region_areas$wcvp_all_area)

# merge with diversity data
alldiv = merge(alldiv, region_areas, by = "species")

# remove temporary objects
rm(wcvp_dist, wcvp_names, region_areas)
```

## Load plant height data, a proxy for body size, which is a proxy for population density
```{r}
height = fread("../workflow/data/eol_plant_height_data_2023-10-03/records.tsv.gz", header = T, fill = T, quote = "", sep = "\t")

# convert feet to meters
height$final_values = height$`Measurement Value`
height$final_values[(height$`Measurement Unit` == "feet")] = height$final_values[(height$`Measurement Unit` == "feet")]*0.3048

# get height measurements for only species in my dataset
height_my_species = NULL
for(i in 1:nrow(alldiv)){
  # print progress
  print(i)
  
  # get species name(s)
  old_name = alldiv$species[i]
  new_name = alldiv$species_updated[i]
  
  # check old name
  height_sub = height[grepl(old_name, height$`Scientific Name`),]
  
  # if new name is different, check new species name
  if(new_name != old_name){
    height_sub = rbind(height_sub, height[grepl(new_name, height$`Scientific Name`),])
  }
  
  # average across values
  height_one_species = data.frame(species = old_name, height = mean(height_sub$final_values, na.rm = T))
  height_my_species = rbind(height_my_species, height_one_species)
}

# add height data to final dataframe
summary(height_my_species$height)
alldiv = merge(alldiv, height_my_species, all.x = T, by = "species")
```

## genome size, ploidy, mating system data
```{r}
# load extra species metadata
metadata = read.table("../results/2023-03-26/species_metadata_2024-04-03.txt", sep = "\t", fill = T, quote = "", header = T)
names(metadata)[1] = "species"

# remove weird space character
metadata$species = gsub("\xa0", " ", metadata$species, useBytes = T)
metadata$species = gsub(" $", "", metadata$species)

# merge metadata with diversity data
alldiv = merge(alldiv, metadata[,c("species", "DNA.amount..1.C..pg.", "Mating.system", "Ploidy", "Height..m.", "cultivation.status", "Generation.time")], by = "species", all.x = T)

# collapse life cycle habit into annual vs not annual
table(alldiv$Generation.time)
alldiv$Generation.time[(alldiv$Generation.time != "annual")] = "perennial.biennial.mixed"
table(alldiv$Generation.time)

# collapse clonal species into outcrossing
# collapse mixed mating into outcrossing
table(alldiv$Mating.system)
alldiv$Mating.system[(alldiv$Mating.system != "outcrossing")] = "selfing.mixed.clonal"
table(alldiv$Mating.system)

# combine different height columns into one
alldiv$height_final = alldiv$height
alldiv$height_final[is.na(alldiv$height_final)] = alldiv$Height..m.[is.na(alldiv$height_final)]

# calculate quantity that's proportional to population size (range/squared height)
alldiv$wcvp_all_popsize = (alldiv$wcvp_all_area * 1e6)/(alldiv$height_final^2)
summary(alldiv$wcvp_all_popsize)

alldiv$wcvp_native_popsize = (alldiv$wcvp_native_area * 1e6)/(alldiv$height_final^2)
summary(alldiv$wcvp_native_popsize)

alldiv$gbif_all_popsize = (alldiv$gbif_all_area * 1e6)/(alldiv$height_final^2)
summary(alldiv$gbif_all_popsize)

alldiv$gbif_native_popsize = (alldiv$gbif_native_area * 1e6)/(alldiv$height_final^2)
summary(alldiv$gbif_native_popsize)

# convert genome size in picograms to megabases
alldiv$genome.size = alldiv$DNA.amount..1.C..pg. * 0.978e9

# calculate average coverage per individual
#alldiv$totalbp[(alldiv$species == "Cicer arietinum")] = 7.5e12 # chickpea was subset
#alldiv$totalbp[(alldiv$species == "Glycine max")] = 7.5e12 # soybean was subset
#alldiv$totalbp[(alldiv$species == "Gossypium hirsutum")] = 7.5e12 # cotton was subset

alldiv$meancov = alldiv$totalbp/(alldiv$genome.size*alldiv$nidv)

# re-order mating system variable
alldiv$Mating.system = factor(alldiv$Mating.system, levels = c("selfing.mixed.clonal", "outcrossing"))
```

## save final organized dataframe
```{r eval=FALSE, include=FALSE}
write.table(alldiv, paste("final_covariates_", today, ".txt", sep = ""), row.names = F, quote = F, sep = "\t")
```

# Initial data analysis

## load saved data for plotting
```{r}
#alldiv2 = read.table("../results/2023-03-26/final_covariates_2024-04-03.txt", sep = "\t", header = T)
alldiv = read.table("../results/2023-03-26/final_covariates_2024-09-08.txt", sep = "\t", header = T)
```

## Prepare phylogenetic tree
```{r}
# write species list to upload to timetree.org, converting some names to those recognized by timetree
timetree_names = alldiv$species
timetree_names[(timetree_names == "Brassica rapa")] = "Brassica rapa subsp. rapa"
timetree_names[(timetree_names == "Brassica oleracea")] = "Brassica oleracea var. capitata"
timetree_names[(timetree_names == "Solanum stenotomum")] = "Solanum tuberosum"
timetree_names[(timetree_names == "Chenopodium quinoa")] = "Chenopodium album"
timetree_names[(timetree_names == "Oryza sativa")] = "Oryza sativa Indica Group"

#write.table(timetree_names, paste("../results/2023-03-26/species_list_", today, ".txt", sep = ""), row.names = F, col.names = F, quote = F)

# load time tree from timetree.org
phy = read.tree("../results/2023-03-26/species_list_2024-04-03.nwk")

# remove underscores
phy$tip.label = gsub("_", " ", phy$tip.label)

# remove variety and subspecies names, convert tip labels to my standard
phy$tip.label = gsub(" var\\..*", "", phy$tip.label)
phy$tip.label = gsub(" subsp\\..*", "", phy$tip.label)

phy$tip.label[(phy$tip.label == "Solanum tuberosum")] = "Solanum stenotomum"
phy$tip.label[(phy$tip.label == "Chenopodium album")] = "Chenopodium quinoa"
phy$tip.label[(phy$tip.label == "Erythranthe guttata")] = "Mimulus guttatus"
phy$tip.label[(phy$tip.label == "Oryza sativa Indica Group")] = "Oryza sativa"

# sanity check
plot(phy)

# check that all species I have data for are in phylogeny and vice versa. Output should be TRUE
all(alldiv$species %in% phy$tip.label)
all(phy$tip.label %in% alldiv$species)
#alldiv$species[!(alldiv$species %in% phy$tip.label)]
```

## plot raw relationships between study variables

```{r}
# scico(2, palette = "lapaz")
# "#190C64" "#FEF2F2"

# collapse long names
alldiv$Generation.time[(alldiv$Generation.time == "perennial.biennial.mixed")] = "not\nannual"
alldiv$Mating.system[(alldiv$Mating.system == "selfing.mixed.clonal")] = "not\noutcrossing"

# check relationship between categorical variables
table(alldiv$Mating.system, alldiv$cultivation.status)
table(alldiv$Mating.system, alldiv$Generation.time)
table(alldiv$Generation.time, alldiv$cultivation.status)

chisq.test(alldiv$Mating.system, alldiv$cultivation.status)
chisq.test(alldiv$Mating.system, alldiv$Generation.time)
chisq.test(alldiv$Generation.time, alldiv$cultivation.status)

# function for simple scatterplot
my_ggplot_wrap <- function(data, xcol, ycol, xname, yname, filename, filew = 5, fileh = 5) {
  
  # get correlation coefficient
  x = with(data, eval(str2expression(xcol)))
  y = with(data, eval(str2expression(ycol)))
  test_results = cor.test(x, y, method = "spearman")
  
  # create label
  cor_label = signif(test_results$estimate, digits = 2)
  p_label = signif(test_results$p.value, digits = 2)
  result_label = paste("\U03C1 = ", cor_label, ", ", "p = ", p_label, sep = "")
  
  # plot
  myplot = ggplot(data = data, aes_string(x = xcol, y = ycol)) + 
    geom_point(pch = 21, color = "black") +
    geom_smooth(method = "lm", color = "purple") +
    theme_classic() +
    #stat_cor(method = "spearman", digits = 2, cor.coef.name = "rho", size = 7) +
    labs(x = xname, y = yname, title = result_label) +
    theme(text = element_text(size = 12))
  
  myplot
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# function to color points by third variable
my_ggplot_wrap_xyz <- function(data, xcol, ycol, xname, yname, filename, zcol, lcol, hcol, lname, mylims, filew = 5, fileh = 5) {
  
  # get correlation coefficient
  x = with(data, eval(str2expression(xcol)))
  y = with(data, eval(str2expression(ycol)))
  test_results = cor.test(x, y, method = "spearman")
  
  # create label
  cor_label = signif(test_results$estimate, digits = 2)
  p_label = signif(test_results$p.value, digits = 2)
  result_label = paste("\U03C1 = ", cor_label, ", ", "p = ", p_label, sep = "")
  
  myplot = ggplot(data = data, aes_string(x = xcol, y = ycol)) + 
    geom_point(aes_string(fill = zcol), pch = 21, color = "black") +
    geom_smooth(method = "loess", color = "purple", n = 20) +
    theme_classic() +
    #stat_cor(method = "spearman", digits = 2, cor.coef.name = "rho", size = 7) +
    scale_fill_gradient(low = lcol, high = hcol, name = lname, limits = mylims) +
    labs(x = xname, y = yname, title = result_label) +
    theme(text = element_text(size = 12))
  
  myplot
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# function to create boxplots
my_ggplot_wrap_box <- function(data, xcol, ycol, xname, yname, filename, filew = 5, fileh = 5) {
  myplot = ggplot(data = data, aes_string(x = xcol, y = ycol)) + 
    geom_boxplot() +
    theme_classic() +
    theme(text = element_text(size = 12)) +
    #stat_kruskal_test(size = 7) +
    #stat_anova_test(type = 3, method = "one_way") +
    labs(x = xname, y = yname)
  
  myplot
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# function to have scatterplot with linear model + 1:1 line
my_ggplot_wrap_line <- function(data, xcol, ycol, xname, yname, filename, filew = 5, fileh = 5) {
  # get correlation coefficient
  x = with(data, eval(str2expression(xcol)))
  y = with(data, eval(str2expression(ycol)))
  test_results = cor.test(x, y, method = "spearman")
  
  # create label
  cor_label = signif(test_results$estimate, digits = 2)
  p_label = signif(test_results$p.value, digits = 2)
  result_label = paste("\U03C1 = ", cor_label, ", ", "p = ", p_label, sep = "")
  
  # combine everything into a plot
  myplot = ggplot(data = data, aes_string(x = xcol, y = ycol)) + 
    geom_point(pch = 21, color = "black") +
    geom_abline(slope = 1, intercept = 0) +
    geom_smooth(method = "lm", color = "purple") + 
    theme_classic() +
    #stat_cor(method = "spearman", digits = 2, cor.coef.name = "rho", size = 7) +
    labs(x = xname, y = yname, title = result_label) +
    theme(text = element_text(size = 12))
  
  myplot
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# function to have scatterplot with linear model only
my_ggplot_wrap_lm <- function(data, xcol, ycol, xname, yname, filename, filew = 5, fileh = 5) {
  # get correlation coefficient
  x = with(data, eval(str2expression(xcol)))
  y = with(data, eval(str2expression(ycol)))
  test_results = cor.test(x, y, method = "spearman")
  
  # create label
  cor_label = signif(test_results$estimate, digits = 2)
  p_label = signif(test_results$p.value, digits = 2)
  result_label = paste("\U03C1 = ", cor_label, ", ", "p = ", p_label, sep = "")
  
  myplot = ggplot(data = data, aes_string(x = xcol, y = ycol)) + 
    geom_point(pch = 21, color = "black") +
    geom_smooth(method = "lm", color = "purple") + 
    theme_classic() +
    #stat_cor(method = "spearman", digits = 2, cor.coef.name = "rho", size = 7) +
    labs(x = xname, y = yname, title = result_label) +
    theme(text = element_text(size = 12))
  
  myplot
  
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# compare two discrete variables
my_ggplot_wrap_barplot = function(data, var1, var2, xname, yname, lname, my_colors, filename, filew = 5, fileh = 5){
    
  # get chi-sq result
  x = with(data, eval(str2expression(var1)))
  y = with(data, eval(str2expression(var2)))
  test_results = chisq.test(x, y)
  
  # construct result label
  stat_label = signif(test_results$statistic, digits = 3)
  p_label = signif(test_results$p.value, digits = 2)
  #result_label = paste("\U03A7^2 = ", stat_label, ", ", "p = ", p_label, sep = "")
  result_label = substitute(paste(chi^2 == s, ", p =", myp), list(s = stat_label, myp = p_label))
  
  myplot = ggplot(alldiv, aes_string(x = var1, fill = var2)) +
    geom_bar(position = "dodge") +
    theme_classic() +
    theme(text = element_text(size = 12)) +
    scale_fill_manual(values = my_colors, name = lname) +
    labs(x = xname, y = yname, title = result_label)

  myplot
  
  ggsave(filename, width = filew, height = fileh)

  return(myplot)
}

# visualize relationship between categorical variables
ida_generation_vs_mating = my_ggplot_wrap_barplot(alldiv, "Generation.time", "Mating.system", "Life cycle habit", "Number of species", "Mating\nsystem", c("#65014B", "#0C4C00"), paste("ida_generation_vs_mating_", today, ".png", sep = ""))

ida_cultivation_vs_mating = my_ggplot_wrap_barplot(alldiv, "Mating.system", "cultivation.status", "Mating system", "Number of species", "Cultivation\nstatus", c("#8C0172", "#B2F2FD"), paste("ida_cultivation_vs_mating_", today, ".png", sep = ""))

ida_cultivation_vs_generation = my_ggplot_wrap_barplot(alldiv, "cultivation.status", "Generation.time", "Cultivation status", "Number of species", "Life cycle\nhabit", c("#9EB0FF", "#FFACAC"), paste("ida_cultivation_vs_generation_", today, ".png", sep = ""))

# scatterplots

## average coverage vs diversity
ida_meancov_vs_pi = my_ggplot_wrap(alldiv, "log10(meancov)", "log10(pi)", "log(mean coverage)", "log(4-fold SNP diversity)", paste("ida_coverage-vs-pi_", today, ".png", sep = ""))

ida_meancov_vs_jac = my_ggplot_wrap(alldiv, "log10(meancov)", "jac", "log(mean coverage)", "Jaccard dissimilarity", paste("ida_coverage-vs-jac_", today, ".png", sep = ""))

ida_meancov_vs_bcd = my_ggplot_wrap(alldiv, "log10(meancov)", "bcd", "log(mean coverage)", "Bray-Curtis dissimilarity", paste("ida_coverage-vs-bcd_", today, ".png", sep = ""))

## depth of sequencing variation vs diversity
ida_cvbp_vs_pi = my_ggplot_wrap(alldiv, "cvbp", "log10(pi)", "coeff. of var. in bp sequenced", "log(4-fold SNP diversity)", paste("ida_cvbp-vs-pi_", today, ".png", sep = ""))

ida_cvbp_vs_jac = my_ggplot_wrap(alldiv, "cvbp", "jac", "coeff. of var. in bp sequenced", "Jaccard dissimilarity", paste("ida_cvbp-vs-jac_", today, ".png", sep = ""))

ida_cvbp_vs_bcd = my_ggplot_wrap(alldiv, "cvbp", "bcd", "coeff. of var. in bp sequenced", "Bray-Curtis dissimilarity", paste("ida_cvbp-vs-bcd_", today, ".png", sep = ""))

## number of individuals sequenced vs diversity
ida_nidv_vs_pi = my_ggplot_wrap(alldiv, "log10(nidv)", "log10(pi)", "log(Number of individuals)", "log(4-fold SNP diversity)", paste("ida_nidv-vs-pi_", today, ".png", sep = ""))

ida_nidv_vs_jac = my_ggplot_wrap(alldiv, "log10(nidv)", "jac", "log(Number of individuals)", "Jaccard dissimilarity", paste("ida_nidv-vs-jac_", today, ".png", sep = ""))

ida_nidv_vs_bcd = my_ggplot_wrap(alldiv, "log10(nidv)", "bcd", "log(Number of individuals)", "Bray-Curtis dissimilarity", paste("ida_nidv-vs-bcd_", today, ".png", sep = ""))

## number of variant sites vs diversity
ida_nvariant_vs_pi = my_ggplot_wrap(alldiv, "log10(nvariant + 1)", "log10(pi)", "log(Number of variant sites + 1)", "log(4-fold SNP diversity)", paste("ida_nvariant-vs-pi_", today, ".png", sep = ""))

# compare diversity, colored by coverage
ida_jac_vs_bcd = my_ggplot_wrap_lm(alldiv, "jac", "bcd", "Jaccard dissimilarity", "Bray-Curtis dissimilarity", paste("ida_jac-vs-bcd_", today, ".png", sep = ""))

ida_pi_vs_jac = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "jac", 'log(4-fold SNP diversity)', "Jaccard dissimilarity", paste("ida_pi-vs-jac_", today, ".png", sep = ""), "log10(meancov)", "#FEF2F2", "#190C64", 'log(mean coverage)', c(min(log10(alldiv$meancov)), max(log10(alldiv$meancov))))

ida_pi_vs_bcd = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "bcd", "log(4-fold SNP diversity)", "Bray-Curtis dissimilarity", paste("ida_pi-vs-bcd_", today, ".png", sep = ""), "log10(meancov)", "#FEF2F2", "#190C64", "log(mean coverage)", c(min(log10(alldiv$meancov)), max(log10(alldiv$meancov))))

# compare diversity measures, colored by variation in coverage
ida_pi_vs_jac_cvbp = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "jac", "log10(4-fold SNP diversity)", "Jaccard dissimilarity", paste("ida_pi-vs-jac_cvbp_", today, ".png", sep = ""), "cvbp", "#FEF2F2", "#190C64", "Coeff.\nof variation\nbp sequenced", c(min(alldiv$cvbp), max(alldiv$cvbp)))

ida_pi_vs_bcd_cvbp = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "bcd", "log10(4-fold SNP diversity)", "Bray-Curtis dissimilarity", paste("ida_pi-vs-bcd_cvbp_", today, ".png", sep = ""), "cvbp", "#FEF2F2", "#190C64", "Coeff.\nof variation\nbp sequenced", c(min(alldiv$cvbp), max(alldiv$cvbp)))

# compare diversity measures, colored by number of individuals sequenced
ida_pi_vs_jac_nidv = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "jac", "log(4-fold SNP diversity)", "Jaccard dissimilarity", paste("ida_pi-vs-jac_nidv_", today, ".png", sep = ""), "log10(nidv)", "#FEF2F2", "#190C64", "log(Number of\nindividuals\nsequenced)", c(min(log10(alldiv$nidv)), max(log10(alldiv$nidv))))

ida_pi_vs_bcd_nidv = my_ggplot_wrap_xyz(alldiv, "log10(pi)", "bcd", "log(4-fold SNP diversity)", "Bray-Curtis dissimilarity", paste("ida_pi-vs-bcd_nidv_", today, ".png", sep = ""), "log10(nidv)", "#FEF2F2", "#190C64", "log(Number of\nindividuals\nsequenced)", c(min(log10(alldiv$nidv)), max(log10(alldiv$nidv))))

# compare k-mer diversity measures
my_ggplot_wrap_xyz(alldiv, "jac", "bcd", "Jaccard dissimilarity", "Bray-Curtis dissimilarity", paste("ida_jac-vs-bcd_", today, ".png", sep = ""), "log10(meancov)", "#FEF2F2", "#190C64", "log10(mean cov/sample)", c(min(log10(alldiv$meancov)), max(log10(alldiv$meancov))))

# compare population size estimates
ida_wcvp_vs_gbif_native = my_ggplot_wrap_line(alldiv, "log10(wcvp_native_area)", "log10(gbif_native_area)", expression("log(WCVP native range)"~km^2), expression("log(GBIF native range)"~km^2), paste("ida_wcvp-vs-gbif_native_", today, ".png", sep = ""))

ida_wcvp_vs_gbif_all = my_ggplot_wrap_line(alldiv, "log10(wcvp_all_area)", "log10(gbif_all_area)", expression("log(WCVP full range)"~km^2), expression("log(GBIF full range)"~km^2), paste("ida_wcvp-vs-gbif_all_", today, ".png", sep = ""))

ida_wcvp_all_vs_wcvp_native = my_ggplot_wrap_line(alldiv, "log10(wcvp_all_area)", "log10(wcvp_native_area)", expression("log(WCVP full range)"~km^2), expression("log(WCVP native range)"~km^2), paste("ida_wcvp-all-vs-wcvp-native_", today, ".png", sep = ""))

ida_gbif_all_vs_gbif_native = my_ggplot_wrap_line(alldiv, "log10(gbif_all_area)", "log10(gbif_native_area)", expression("log(GBIF full range)"~km^2), expression("log(GBIF native range)"~km^2), paste("ida_gbif-all-vs-gbif-native_", today, ".png", sep = ""))

# plot diversity by mating system
ida_mating_vs_pi = my_ggplot_wrap_box(alldiv, "Mating.system", "log10(pi)", "Mating system", "log10(4-fold SNP diversity)", paste("mating_vs_pi_", today, ".png", sep = ""))

ida_mating_vs_jac = my_ggplot_wrap_box(alldiv, "Mating.system", "jac", "Mating system", "Jaccard dissimilarity", paste("mating_vs_jac_", today, ".png", sep = ""))

ida_mating_vs_bcd = my_ggplot_wrap_box(alldiv, "Mating.system", "bcd", "Mating system", "Bray-Curtis dissimilarity", paste("mating_vs_bcd_", today, ".png", sep = ""))

# plot diversity by ploidy
ida_ploidy_vs_pi = my_ggplot_wrap_box(alldiv, "Ploidy", "log10(pi)", "Ploidy", "log10(4-fold SNP diversity)", paste("ploidy_vs_pi_", today, ".png", sep = ""))

ida_ploidy_vs_jac = my_ggplot_wrap_box(alldiv, "Ploidy", "jac", "Ploidy", "Jaccard dissimilarity", paste("ploidy_vs_jac_", today, ".png", sep = ""))

ida_ploidy_vs_bcd = my_ggplot_wrap_box(alldiv, "Ploidy", "bcd", "Ploidy", "Bray-Curtis dissimilarity", paste("ploidy_vs_bcd_", today, ".png", sep = ""))

# plot diversity by cultivation status
ida_cult_vs_pi = my_ggplot_wrap_box(alldiv, "cultivation.status", "log10(pi)", "Cultivation status", "log10(4-fold SNP diversity)", paste("cultivation_vs_pi_", today, ".png", sep = ""))

ida_cult_vs_jac = my_ggplot_wrap_box(alldiv, "cultivation.status", "jac", "Cultivation status", "Jaccard dissimilarity", paste("cultivation_vs_jac_", today, ".png", sep = ""))

ida_cult_vs_bcd = my_ggplot_wrap_box(alldiv, "cultivation.status", "bcd", "Cultivation status", "Bray-Curtis dissimilarity", paste("cultivation_vs_bcd_", today, ".png", sep = ""))

# plot height for each species, from tallest to shortest
#species_order = alldiv$species[sort(alldiv$height_final, decreasing = T, index.return = T)$ix]

#ggplot(alldiv, aes(x = species, y = height_final)) +
#  geom_bar(stat = "identity") +
#  theme_classic() +
#  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
#  scale_x_discrete(limits = species_order) + 
#  labs(x = "Species", y = "Height (m)")

#ggsave(paste("ida_height-by-species_", today, ".png", sep = ""), width = 16, height = 4.5, units = "in")
```

# Confirmatory data analysis

## remove outliers before analysis
```{r}
# For now, remove species where invariant sites were not properly called
#alldiv = alldiv[(alldiv$ninvariant > 0),]

# For now, remove species where variants sites were not properly called
#alldiv = alldiv[(alldiv$nvariant > 100),]

# remove critically endangered species
# alldiv = alldiv[(alldiv$rangeArea > 10),]

# remove species where there aren't many sampled individuals
#alldiv = alldiv[(alldiv$nidv >= 5),]

# remove any rows that are NA for some reason
#alldiv = alldiv[!is.na(alldiv$species),]

# remove all species with <1.5x coverage on average across individuals
#subdiv = alldiv[(log10(alldiv$meancov) > log10(1.5)),]
#subdiv = alldiv[(log10(alldiv$meancov) > log10(1.5) | alldiv$nvariant >= 1000),]
subdiv = alldiv[(alldiv$nvariant >= 1000 & alldiv$meancov > 0.5),]
#subdiv = subdiv[(subdiv$species != "Brassica rapa"),] # omit other outlier species
```

## replot trends
```{r}
my_ggplot_wrap(subdiv, "log10(meancov)", "log10(pi)", "log10(mean coverage)", "log10(4-fold SNP diversity)", paste("cda_coverage-vs-pi_", today, ".png", sep = ""))

my_ggplot_wrap(subdiv, "log10(meancov)", "jac", "log10(mean coverage)", "Jaccard dissimilarity", paste("cda_coverage-vs-jac_", today, ".png", sep = ""))

my_ggplot_wrap(subdiv, "log10(meancov)", "bcd", "log10(mean coverage)", "Bray-Curtis dissimilarity", paste("cda_coverage-vs-bcd_", today, ".png", sep = ""))

my_ggplot_wrap(subdiv, "cvbp", "log10(pi)", "coeff. of var. in bp sequenced", "log10(4-fold SNP diversity)", paste("cda_cvbp-vs-pi_", today, ".png", sep = ""))

my_ggplot_wrap(subdiv, "cvbp", "jac", "coeff. of var. in bp sequenced", "Jaccard dissimilarity", paste("cda_cvbp-vs-jac_", today, ".png", sep = ""))

my_ggplot_wrap(subdiv, "cvbp", "bcd", "coeff. of var. in bp sequenced", "Bray-Curtis dissimilarity", paste("cda_cvbp-vs-bcd_", today, ".png", sep = ""))

cda_pi_vs_jac = my_ggplot_wrap_xyz(subdiv, "log10(pi)", "jac", "log(4-fold SNP diversity)", "Jaccard dissimilarity", paste("cda_pi-vs-jac_", today, ".png", sep = ""), "log10(meancov)", "#FEF2F2", "#190C64", "log(mean coverage)", c(min(log10(alldiv$meancov)), max(log10(alldiv$meancov))))

cda_pi_vs_bcd = my_ggplot_wrap_xyz(subdiv, "log10(pi)", "bcd", "log(4-fold SNP diversity)", "Bray-Curtis dissimilarity", paste("cda_pi-vs-bcd_", today, ".png", sep = ""), "log10(meancov)", "#FEF2F2", "#190C64", "log(mean coverage)", c(min(log10(alldiv$meancov)), max(log10(alldiv$meancov))))

my_ggplot_wrap(subdiv, "jac", "bcd", "Jaccard dissimilarity", "Bray-Curtis dissimilarity", paste("cda_jac-vs-bcd_", today, ".png", sep = ""))

# compare population size proxies with diversity
cda_pi_vs_wcvp = my_ggplot_wrap(subdiv, "log10(pi)", "log10(wcvp_native_popsize)", "log(4-fold SNP diversity)", "log(WCVP native ratio)", paste("cda_pi-vs-wcvp_", today, ".png", sep = ""))

cda_pi_vs_gbif = my_ggplot_wrap(subdiv, "log10(pi)", "log10(gbif_native_popsize)", "log(4-fold SNP diversity)", "log(GBIF native ratio)", paste("cda_pi-vs-gbif_", today, ".png", sep = ""))

cda_jac_vs_wcvp = my_ggplot_wrap(subdiv, "jac", "log10(wcvp_native_popsize)", "Jaccard dissimilarity", "log(WCVP native ratio)", paste("cda_jac-vs-wcvp_", today, ".png", sep = ""))

cda_jac_vs_gbif = my_ggplot_wrap(subdiv, "jac", "log10(gbif_native_popsize)", "Jaccard dissimilarity", "log(GBIF native ratio)", paste("cda_jac-vs-gbif_", today, ".png", sep = ""))

cda_bcd_vs_wcvp = my_ggplot_wrap(subdiv, "bcd", "log10(wcvp_native_popsize)", "Bray-Curtis dissimilarity", "log(WCVP native ratio)", paste("cda_bcd-vs-wcvp_", today, ".png", sep = ""))

cda_bcd_vs_gbif = my_ggplot_wrap(subdiv, "bcd", "log10(gbif_native_popsize)", "Bray-Curtis dissimilarity", "log10(GBIF native ratio)", paste("cda_bcd-vs-gbif_", today, ".png", sep = ""))
```

## Combine tree and data to get comparative object
```{r}
# create comparative data object for caper models
# keep GBIF data separate from WCVP because GBIF might have missing values
compdata = comparative.data(phy, subdiv[,c("species","pi", "jac", "bcd", "genome.size", "meancov", "nidv", "cvbp", "cultivation.status", "Mating.system", "Generation.time","wcvp_native_popsize", "wcvp_all_popsize", "wcvp_native_area", "wcvp_all_area", "gbif_native_popsize", "gbif_all_popsize", "gbif_native_area", "gbif_all_area", "height_final")], names.col = "species", vcv = T)

#gbif_compdata = comparative.data(phy, subdiv[,c("species","pi", "jac", "bcd", "genome.size", "meancov", "nidv", "cvbp", "cultivation.status", "Mating.system", "gbif_native_popsize", "gbif_all_popsize", "gbif_native_area", "gbif_all_area")], names.col = "species", vcv = T)
```

## fit models

### population size vs diversity

```{r}
# check for multicollinarity before fitting models
#cor(wcvp_compdata$data[,c(-1, -8, -9)], method = "spearman", use = "pairwise.complete.obs")

##############################WCVP native population size vs diversity###################
mod_wcvp_native_popsize_pi = pgls(scale(log10(pi)) ~ log10(wcvp_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_native_popsize_pi)
#plot(mod_wcvp_native_popsize_pi)

mod_wcvp_native_popsize_jac = pgls(scale(jac) ~ log10(wcvp_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_native_popsize_jac)
#plot(mod_wcvp_native_popsize_jac)

mod_wcvp_native_popsize_bcd = pgls(scale(bcd) ~ log10(wcvp_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_native_popsize_bcd)
#plot(mod_wcvp_native_popsize_bcd)

##############################WCVP all population size vs diversity######################

mod_wcvp_all_popsize_pi = pgls(scale(log10(pi)) ~ log10(wcvp_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_all_popsize_pi)
#plot(mod_wcvp_all_popsize_pi)

mod_wcvp_all_popsize_jac = pgls(scale(jac) ~ log10(wcvp_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_all_popsize_jac)
#plot(mod_wcvp_all_popsize_jac)

mod_wcvp_all_popsize_bcd = pgls(scale(bcd) ~ log10(wcvp_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_wcvp_all_popsize_bcd)
#plot(mod_wcvp_all_popsize_bcd)

############################GBIF native population size vs diversity####################

mod_gbif_native_popsize_pi = pgls(scale(log10(pi)) ~ log10(gbif_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_native_popsize_pi)

mod_gbif_native_popsize_jac = pgls(scale(jac) ~ log10(gbif_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_native_popsize_jac)

mod_gbif_native_popsize_bcd = pgls(scale(bcd) ~ log10(gbif_native_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_native_popsize_bcd)

##########################GBIF all population size vs diversity##########################

mod_gbif_all_popsize_pi = pgls(scale(log10(pi)) ~ log10(gbif_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_all_popsize_pi)

mod_gbif_all_popsize_jac = pgls(scale(jac) ~ log10(gbif_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_all_popsize_jac)

mod_gbif_all_popsize_bcd = pgls(scale(bcd) ~ log10(gbif_all_popsize) + Mating.system + cultivation.status + Generation.time, compdata)
#summary(mod_gbif_all_popsize_bcd)

####WCVP all range size vs diversity####
mod_wcvp_all_area_pi = pgls(scale(log10(pi)) ~ log10(wcvp_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_wcvp_all_area_jac = pgls(scale(jac) ~ log10(wcvp_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_wcvp_all_area_bcd = pgls(scale(bcd) ~ log10(wcvp_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

####WCVP native range size vs diversity####
mod_wcvp_native_area_pi = pgls(scale(log10(pi)) ~ log10(wcvp_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_wcvp_native_area_jac = pgls(scale(jac) ~ log10(wcvp_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_wcvp_native_area_bcd = pgls(scale(bcd) ~ log10(wcvp_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

####GBIF all range size vs diversity####
mod_gbif_all_area_pi = pgls(scale(log10(pi)) ~ log10(gbif_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_gbif_all_area_jac = pgls(scale(jac) ~ log10(gbif_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_gbif_all_area_bcd = pgls(scale(bcd) ~ log10(gbif_all_area) + Mating.system + cultivation.status + Generation.time, compdata)

####GBIF native range size vs diversity####
mod_gbif_native_area_pi = pgls(scale(log10(pi)) ~ log10(gbif_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_gbif_native_area_jac = pgls(scale(jac) ~ log10(gbif_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

mod_gbif_native_area_bcd = pgls(scale(bcd) ~ log10(gbif_native_area) + Mating.system + cultivation.status + Generation.time, compdata)

####plant height vs diversity####
mod_height_pi = pgls(scale(log10(pi)) ~ log10(height_final) + Mating.system + cultivation.status + Generation.time, compdata)

mod_height_jac = pgls(scale(jac) ~ log10(height_final) + Mating.system + cultivation.status + Generation.time, compdata)

mod_height_bcd = pgls(scale(bcd) ~ log10(height_final) + Mating.system + cultivation.status + Generation.time, compdata)

# create model matrices for plotting multiple regressions
mod_mat_wcvp_native = model.matrix(~ log10(wcvp_native_popsize) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_wcvp_all = model.matrix(~ log10(wcvp_all_popsize) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_gbif_native = model.matrix(~ log10(gbif_native_popsize) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_gbif_all = model.matrix(~ log10(gbif_all_popsize) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_wcvp_native_area = model.matrix(~ log10(wcvp_native_area) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_wcvp_all_area = model.matrix(~ log10(wcvp_all_area) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_gbif_native_area = model.matrix(~ log10(gbif_native_area) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_gbif_all_area = model.matrix(~ log10(gbif_all_area) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

mod_mat_height = model.matrix(~ log10(height_final) + Mating.system + cultivation.status + Generation.time, data = compdata$data)

####Plot multiple regressions####

# function to perform phylogenetic least squares, adjust for covariates, and plot primary variable against responses
correct_and_plot = function(mod_mat, y, V, primary, xlab, ylab, xpos, ypos){
  # Get cholesky decomposition of variance covariance matrix
  # then invert the matrix
  # then transpose it
  C_inv = t(solve(chol(V)))
  
  # scale data by phylogenetic effect
  # This translates the generalized least squares model into an ordinary least squares model
  y_star = C_inv %*% y
  X_star = C_inv %*% mod_mat
  
  # perform ordinary least squares regression
  # omit intercept because it's included in design matrix
  mod_star = lm(y_star ~ X_star - 1)
  coeff_star = coefficients(mod_star)
  
  names(coeff_star) = gsub("^X_star", "", names(coeff_star)) # remove new names from coefficients
  
  # extract p-value for slope of primary variable
  pvalues = summary(mod_star)$coefficients
  rownames(pvalues) = gsub("^X_star", "", rownames(pvalues)) # remove new names from coefficients
  slope_value = signif(pvalues[primary,1], digits = 2) # get slope
  slope_se_value = signif(pvalues[primary,2], digits = 2) # get standard error of slope
  slope_p_value = signif(pvalues[primary,4], digits = 2) # get p-value for slope
  
  # correct response for phylogeny and other covariates, then plot against predictor of interest
  technicals = colnames(X_star)[(colnames(X_star) != primary)]
  y_star_corrected = y_star - (X_star[,technicals] %*% coeff_star[technicals])
  
  # Create label that will go on plot with lm information
  modlabel = paste("slope=", slope_value, "+/-", slope_se_value, "\n", "p = ", slope_p_value, "\n", sep = "")
  
  # plot result
  myplot = ggplot(mapping = aes(x = X_star[,primary], y = y_star_corrected)) +
    geom_point(pch = 21, color = "black") +
    geom_abline(intercept = 0, 
                slope = coeff_star[primary] - slope_se_value, 
                color = "black", 
                linetype = "dotted") +
    geom_abline(intercept = 0, 
                slope = coeff_star[primary], 
                color = "purple") +
    geom_abline(intercept = 0, 
                slope = coeff_star[primary] + slope_se_value, 
                color = "black", 
                linetype = "dotted") +
    theme_classic() +
    theme(text = element_text(size = 10)) +
    #stat_regline_equation(aes(label = paste(..eq.label.., ..adj.rr.label.., sep = "~~~~")), size = 5) +
    #annotate(geom = "text", x = xpos, y = ypos, label = modlabel, size = 7) +
    ggtitle(modlabel) +
    labs(x = xlab, y = ylab)
  
  #print(X_star[,primary])
    
  return(myplot)
}

# create plots, save to objects

## wcvp

wcvp_native_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_native, mod_wcvp_native_popsize_pi$y, mod_wcvp_native_popsize_pi$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

wcvp_all_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_all, mod_wcvp_all_popsize_pi$y, mod_wcvp_all_popsize_pi$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

wcvp_all_area_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_all_area, mod_wcvp_all_area_pi$y, mod_wcvp_all_area_pi$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

wcvp_native_area_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_native_area, mod_wcvp_native_area_pi$y, mod_wcvp_native_area_pi$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

wcvp_native_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_native, mod_wcvp_native_popsize_jac$y, mod_wcvp_native_popsize_jac$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

wcvp_all_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_all, mod_wcvp_all_popsize_jac$y, mod_wcvp_all_popsize_jac$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

wcvp_all_area_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_all_area, mod_wcvp_all_area_jac$y, mod_wcvp_all_area_jac$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

wcvp_native_area_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_native_area, mod_wcvp_native_area_jac$y, mod_wcvp_native_area_jac$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

wcvp_native_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_native, mod_wcvp_native_popsize_bcd$y, mod_wcvp_native_popsize_bcd$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

wcvp_all_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_all, mod_wcvp_all_popsize_bcd$y, mod_wcvp_all_popsize_bcd$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

wcvp_all_area_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_all_area, mod_wcvp_all_area_bcd$y, mod_wcvp_all_area_bcd$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

wcvp_native_area_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_native_area, mod_wcvp_native_area_bcd$y, mod_wcvp_native_area_bcd$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif

# pi

gbif_native_vs_pi_corrected = correct_and_plot(mod_mat_gbif_native, mod_gbif_native_popsize_pi$y, mod_gbif_native_popsize_pi$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

gbif_all_vs_pi_corrected = correct_and_plot(mod_mat_gbif_all, mod_gbif_all_popsize_pi$y, mod_gbif_all_popsize_pi$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

gbif_all_area_vs_pi_corrected = correct_and_plot(mod_mat_gbif_all_area, mod_gbif_all_area_pi$y, mod_gbif_all_area_pi$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")
  
gbif_native_area_vs_pi_corrected = correct_and_plot(mod_mat_gbif_native_area, mod_gbif_native_area_pi$y, mod_gbif_native_area_pi$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

# JAC

gbif_native_vs_jac_corrected = correct_and_plot(mod_mat_gbif_native, mod_gbif_native_popsize_jac$y, mod_gbif_native_popsize_jac$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

gbif_all_vs_jac_corrected = correct_and_plot(mod_mat_gbif_all, mod_gbif_all_popsize_jac$y, mod_gbif_all_popsize_jac$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

gbif_all_area_vs_jac_corrected = correct_and_plot(mod_mat_gbif_all_area, mod_gbif_all_area_jac$y, mod_gbif_all_area_jac$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")
  
gbif_native_area_vs_jac_corrected = correct_and_plot(mod_mat_gbif_native_area, mod_gbif_native_area_jac$y, mod_gbif_native_area_jac$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

# bcd

gbif_native_vs_bcd_corrected = correct_and_plot(mod_mat_gbif_native, mod_gbif_native_popsize_bcd$y, mod_gbif_native_popsize_bcd$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

gbif_all_vs_bcd_corrected = correct_and_plot(mod_mat_gbif_all, mod_gbif_all_popsize_bcd$y, mod_gbif_all_popsize_bcd$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

gbif_all_area_vs_bcd_corrected = correct_and_plot(mod_mat_gbif_all_area, mod_gbif_all_area_bcd$y, mod_gbif_all_area_bcd$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")
  
gbif_native_area_vs_bcd_corrected = correct_and_plot(mod_mat_gbif_native_area, mod_gbif_native_area_bcd$y, mod_gbif_native_area_bcd$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## height

height_vs_pi_corrected = correct_and_plot(mod_mat_height, mod_height_pi$y, mod_height_pi$Vt, "log10(height_final)", "log(height, m)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

height_vs_jac_corrected = correct_and_plot(mod_mat_height, mod_height_jac$y, mod_height_jac$Vt, "log10(height_final)", "log(height, m)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

height_vs_bcd_corrected = correct_and_plot(mod_mat_height, mod_height_bcd$y, mod_height_bcd$Vt, "log10(height_final)", "log(height, m)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")


#wcvp_pop_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_pop_vs_pi, mod_pi_1$y, mod_pi_1$Vt, "log10(wcvp_native_popsize)", "log10(WCVP pop size ratio)\nscaled", "log10(4-fold SNP diversity)\nscaled and adjusted")

#wcvp_pop_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_pop_vs_jac, mod_jac_1$y, mod_jac_1$Vt, "log10(wcvp_native_popsize)", "log10(WCVP pop size ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

#wcvp_pop_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_pop_vs_bcd, mod_bcd_1$y, mod_bcd_1$Vt, "log10(wcvp_native_popsize)", "log10(WCVP pop size ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

#wcvp_area_vs_pi_corrected = correct_and_plot(mod_mat_wcvp_area_vs_pi, mod_pi_wcvp_area_3$y, mod_pi_wcvp_area_3$Vt, "log10(area)", "log10(WCVP range in square km)\nscaled", "log10(4-fold SNP diversity)\nscaled and adjusted")

#wcvp_area_vs_jac_corrected = correct_and_plot(mod_mat_wcvp_area_vs_jac, mod_jac_wcvp_area_1$y, mod_jac_wcvp_area_1$Vt, "log10(area)", "log10(WCVP range in square km)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

#wcvp_area_vs_bcd_corrected = correct_and_plot(mod_mat_wcvp_area_vs_bcd, mod_bcd_wcvp_area_2$y, mod_bcd_wcvp_area_2$Vt, "log10(area)", "log10(WCVP range in square km)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

```

### Genome size vs diversity
```{r}
# WCVP

## native + invaded ratio
genome_wcvp_all_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(wcvp_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_vs_pi)
#plot(genome_wcvp_all_vs_pi)

genome_wcvp_all_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(wcvp_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_vs_jac)

genome_wcvp_all_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(wcvp_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_vs_bcd)

genome_mod_mat_wcvp_all = model.matrix(~ log10(genome.size) + log10(wcvp_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native only ratio
genome_wcvp_native_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(wcvp_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_vs_pi)

genome_wcvp_native_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(wcvp_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_vs_jac)

genome_wcvp_native_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(wcvp_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_vs_bcd)

genome_mod_mat_wcvp_native = model.matrix(~ log10(genome.size) + log10(wcvp_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native + invaded range size
genome_wcvp_all_area_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(wcvp_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_area_vs_pi)

genome_wcvp_all_area_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(wcvp_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_area_vs_jac)

genome_wcvp_all_area_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(wcvp_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_all_area_vs_bcd)

genome_mod_mat_wcvp_all_area = model.matrix(~ log10(genome.size) + log10(wcvp_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native range size
genome_wcvp_native_area_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(wcvp_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_area_vs_pi)

genome_wcvp_native_area_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(wcvp_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_area_vs_jac)

genome_wcvp_native_area_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(wcvp_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_wcvp_native_area_vs_bcd)

genome_mod_mat_wcvp_native_area = model.matrix(~ log10(genome.size) + log10(wcvp_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

# GBIF

## native + invaded ratio
genome_gbif_all_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(gbif_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_vs_pi)

genome_gbif_all_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(gbif_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_vs_jac)

genome_gbif_all_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(gbif_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_vs_bcd)

genome_mod_mat_gbif_all = model.matrix(~ log10(genome.size) + log10(gbif_all_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native only ratio
genome_gbif_native_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(gbif_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_vs_pi)

genome_gbif_native_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(gbif_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_vs_jac)

genome_gbif_native_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(gbif_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_vs_bcd)

genome_mod_mat_gbif_native = model.matrix(~ log10(genome.size) + log10(gbif_native_popsize) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native + invaded range size
genome_gbif_all_area_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(gbif_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_area_vs_pi)

genome_gbif_all_area_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(gbif_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_area_vs_jac)

genome_gbif_all_area_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(gbif_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_all_area_vs_bcd)

genome_mod_mat_gbif_all_area = model.matrix(~ log10(genome.size) + log10(gbif_all_area) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

## native range size
genome_gbif_native_area_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(gbif_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_area_vs_pi)

genome_gbif_native_area_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(gbif_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_area_vs_jac)

genome_gbif_native_area_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(gbif_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_gbif_native_area_vs_bcd)

genome_mod_mat_gbif_native_area = model.matrix(~ log10(genome.size) + log10(gbif_native_area) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

# height
genome_height_vs_pi = pgls(scale(log10(pi)) ~ log10(genome.size) + log10(height_final) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_height_vs_pi)

genome_height_vs_jac = pgls(scale(jac) ~ log10(genome.size) + log10(height_final) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_height_vs_jac)

genome_height_vs_bcd = pgls(scale(bcd) ~ log10(genome.size) + log10(height_final) + Mating.system + Generation.time + cultivation.status, data = compdata)
#summary(genome_height_vs_bcd)

genome_mod_mat_height = model.matrix(~ log10(genome.size) + log10(height_final) + Mating.system + Generation.time + cultivation.status, data = compdata$data)

# create plots

## gbif all ratio
genome_gbif_all_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_pi$y, genome_gbif_all_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_gbif_all_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_jac$y, genome_gbif_all_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_gbif_all_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_bcd$y, genome_gbif_all_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif native ratio
genome_gbif_native_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_pi$y, genome_gbif_native_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_gbif_native_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_jac$y, genome_gbif_native_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_gbif_native_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_bcd$y, genome_gbif_native_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif all range size
genome_gbif_all_area_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_pi$y, genome_gbif_all_area_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_gbif_all_area_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_jac$y, genome_gbif_all_area_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_gbif_all_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_bcd$y, genome_gbif_all_area_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif native range size
genome_gbif_native_area_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_pi$y, genome_gbif_native_area_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_gbif_native_area_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_jac$y, genome_gbif_native_area_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_gbif_native_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_bcd$y, genome_gbif_native_area_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp all
genome_wcvp_all_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_pi$y, genome_wcvp_all_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_wcvp_all_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_jac$y, genome_wcvp_all_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_wcvp_all_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_bcd$y, genome_wcvp_all_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp native
genome_wcvp_native_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_pi$y, genome_wcvp_native_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_wcvp_native_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_jac$y, genome_wcvp_native_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_wcvp_native_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_bcd$y, genome_wcvp_native_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp all range size
genome_wcvp_all_area_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_pi$y, genome_wcvp_all_area_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_wcvp_all_area_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_jac$y, genome_wcvp_all_area_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_wcvp_all_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_bcd$y, genome_wcvp_all_area_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp native range size
genome_wcvp_native_area_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_pi$y, genome_wcvp_native_area_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_wcvp_native_area_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_jac$y, genome_wcvp_native_area_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_wcvp_native_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_bcd$y, genome_wcvp_native_area_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## height
genome_height_vs_pi_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_pi$y, genome_height_vs_pi$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

genome_height_vs_jac_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_jac$y, genome_height_vs_jac$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

genome_height_vs_bcd_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_bcd$y, genome_height_vs_bcd$Vt, "log10(genome.size)", "log(genome size in bp)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")
```

### Population size vs diversity, controlling for genome size
```{r}
## gbif all ratio
pop_genome_gbif_all_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_pi$y, genome_gbif_all_vs_pi$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_gbif_all_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_jac$y, genome_gbif_all_vs_jac$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_gbif_all_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_all, genome_gbif_all_vs_bcd$y, genome_gbif_all_vs_bcd$Vt, "log10(gbif_all_popsize)", "log(GBIF full ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif native ratio
pop_genome_gbif_native_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_pi$y, genome_gbif_native_vs_pi$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_gbif_native_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_jac$y, genome_gbif_native_vs_jac$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_gbif_native_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_native, genome_gbif_native_vs_bcd$y, genome_gbif_native_vs_bcd$Vt, "log10(gbif_native_popsize)", "log(GBIF native ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif all range size
pop_genome_gbif_all_area_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_pi$y, genome_gbif_all_area_vs_pi$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_gbif_all_area_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_jac$y, genome_gbif_all_area_vs_jac$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_gbif_all_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_all_area, genome_gbif_all_area_vs_bcd$y, genome_gbif_all_area_vs_bcd$Vt, "log10(gbif_all_area)", "log(GBIF full range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## gbif native range size
pop_genome_gbif_native_area_vs_pi_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_pi$y, genome_gbif_native_area_vs_pi$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_gbif_native_area_vs_jac_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_jac$y, genome_gbif_native_area_vs_jac$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_gbif_native_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_gbif_native_area, genome_gbif_native_area_vs_bcd$y, genome_gbif_native_area_vs_bcd$Vt, "log10(gbif_native_area)", "log(GBIF native range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp all
pop_genome_wcvp_all_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_pi$y, genome_wcvp_all_vs_pi$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_wcvp_all_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_jac$y, genome_wcvp_all_vs_jac$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_wcvp_all_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_all, genome_wcvp_all_vs_bcd$y, genome_wcvp_all_vs_bcd$Vt, "log10(wcvp_all_popsize)", "log(WCVP full ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp native
pop_genome_wcvp_native_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_pi$y, genome_wcvp_native_vs_pi$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_wcvp_native_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_jac$y, genome_wcvp_native_vs_jac$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_wcvp_native_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_native, genome_wcvp_native_vs_bcd$y, genome_wcvp_native_vs_bcd$Vt, "log10(wcvp_native_popsize)", "log(WCVP native ratio)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp all range size
pop_genome_wcvp_all_area_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_pi$y, genome_wcvp_all_area_vs_pi$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_wcvp_all_area_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_jac$y, genome_wcvp_all_area_vs_jac$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_wcvp_all_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_all_area, genome_wcvp_all_area_vs_bcd$y, genome_wcvp_all_area_vs_bcd$Vt, "log10(wcvp_all_area)", "log(WCVP full range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## wcvp native range size
pop_genome_wcvp_native_area_vs_pi_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_pi$y, genome_wcvp_native_area_vs_pi$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_wcvp_native_area_vs_jac_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_jac$y, genome_wcvp_native_area_vs_jac$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_wcvp_native_area_vs_bcd_figure = correct_and_plot(genome_mod_mat_wcvp_native_area, genome_wcvp_native_area_vs_bcd$y, genome_wcvp_native_area_vs_bcd$Vt, "log10(wcvp_native_area)", "log(WCVP native range, km^2)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")

## height
pop_genome_height_vs_pi_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_pi$y, genome_height_vs_pi$Vt, "log10(height_final)", "log(height in m)\nscaled", "log(4-fold SNP diversity)\nscaled and adjusted")

pop_genome_height_vs_jac_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_jac$y, genome_height_vs_jac$Vt, "log10(height_final)", "log(height in m)\nscaled", "Jaccard dissimilarity\nscaled and adjusted")

pop_genome_height_vs_bcd_figure = correct_and_plot(genome_mod_mat_height, genome_height_vs_bcd$y, genome_height_vs_bcd$Vt, "log10(height_final)", "log(height in m)\nscaled", "Bray-Curtis dissimilarity\nscaled and adjusted")
```


# Final plots

```{r}
# make initial data analysis plots, confirming relationship between WCVP and GBIF, jaccard and bray-curtis
ggarrange(ida_wcvp_vs_gbif_all, ida_wcvp_vs_gbif_native, ida_wcvp_all_vs_wcvp_native, ida_gbif_all_vs_gbif_native, nrow = 2, ncol = 2, labels = c("A", "B", "C", "D"), font.label = list(size = 12), hjust = -2)

#ggsave(paste("wcvp_vs_gbif_", today, ".tiff", sep = ""), width = 6, height = 6, scale = 1, compression = "lzw")
ggsave(paste("wcvp_vs_gbif_", today, ".jpg", sep = ""), width = 6, height = 6, dpi = 350)

# Jaccard vs bray-curtis dissimilarity
ggsave(paste("ida_jac_vs_bcd_", today, ".jpg", sep = ""), ida_jac_vs_bcd, width = 4, height = 4, dpi = 350)

# Look at sequencing metrics influence diversity
ggarrange(ida_meancov_vs_pi, ida_cvbp_vs_pi, ida_nidv_vs_pi, ida_meancov_vs_jac, ida_cvbp_vs_jac, ida_nidv_vs_jac, ida_meancov_vs_bcd, ida_cvbp_vs_bcd, ida_nidv_vs_bcd, ncol = 3, nrow = 3, labels = c("A", "B", "C", "D", "E", "F", "G", "H", "I"), font.label = list(size = 12))

ggsave(paste("coverage_vs_diversity_", today, ".jpg", sep = ""), width = 10, height = 8, units = "in", dpi = 350)

# Look at how mating system, cultivation status, ploidy influence diversity
#ggarrange(ida_mating_vs_pi, ida_cult_vs_pi, ida_ploidy_vs_pi, ida_mating_vs_jac, ida_cult_vs_jac, ida_ploidy_vs_jac, ida_mating_vs_bcd, ida_cult_vs_bcd, ida_ploidy_vs_bcd, ncol = 3, nrow = 3, labels = c("A", "B", "C", "D", "E", "F", "G", "H"))

#ggsave(paste("mating_cultivation_ploidy_vs_diversity_", today, ".jpg", sep = ""), width = 10, height = 10, units = "in", dpi = 350)

# Compare how k-mer metrics compare to SNP metrics after removing low coverage data
ggarrange(ida_pi_vs_bcd, cda_pi_vs_bcd, ncol = 2, nrow = 1, common.legend = T, labels = c("A", "B"), legend = "top", font.label = list(size = 12))

ggsave(paste("pi_vs_bcd_high_vs_low_cov_", today, ".tiff", sep = ""), width = 6, height = 4, units = "in", compression = "lzw")
ggsave(paste("pi_vs_bcd_high_vs_low_cov_", today, ".jpg", sep = ""), width = 6, height = 4, units = "in", dpi = 350)

ggarrange(ida_pi_vs_jac, cda_pi_vs_jac, ncol = 2, nrow = 1, common.legend = T, labels = c("A", "B"), legend = "top", font.label = list(size = 12))

ggsave(paste("pi_vs_jac_high_vs_low_cov_", today, ".jpg", sep = ""), width = 6, height = 4, units = "in", dpi = 350)

# Look at how cv in sequencing affect SNP vs k-mer comparison
ggarrange(ida_pi_vs_jac_cvbp, ida_pi_vs_bcd_cvbp, ncol = 2, nrow = 1, common.legend = T, labels = c("A", "B"), legend = "top", font.label = list(size = 12))

ggsave(paste("pi_vs_jac_and_bcd_cvbp_", today, ".jpg", sep = ""), width = 9, height = 6, units = "in", dpi = 350)

# Look at how number of individuals sequenced affects SNP vs k-mer comparison
ggarrange(ida_pi_vs_jac_nidv, ida_pi_vs_bcd_nidv, ncol = 2, nrow = 1, common.legend = T, labels = c("A", "B"), legend = "top", font.label = list(size = 12))

ggsave(paste("pi_vs_jac_and_bcd_nidv_", today, ".jpg", sep = ""), width = 9, height = 6, units = "in", dpi = 350)

# Simple barplot of how cultivation varies with generation and mating system
ggarrange(ida_cultivation_vs_generation, ida_cultivation_vs_mating, ida_generation_vs_mating, labels = c("A", "B", "C"), font.label = list(size = 12), ncol = 3, nrow = 1, legend = "right")

ggsave(paste("cultivation_vs_generation_and_mating", today, ".jpg", sep = ""), width = 15, height = 4, units = "in", dpi = 350)

# Compare population sizes to diversity metrics, without correcting for phylogeny
ggarrange(cda_pi_vs_gbif, cda_pi_vs_wcvp, cda_jac_vs_gbif, cda_jac_vs_wcvp, cda_bcd_vs_gbif, cda_bcd_vs_wcvp, nrow = 3, ncol = 2, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("diversity_vs_popsize_", today, ".jpg", sep = ""), width = 6, height = 8, units = "in", dpi = 350)

#### plot phylogenetic least squares models ####

## population size vs diversity, excluding genome size

### Figure 3
ggarrange(wcvp_native_vs_pi_corrected, wcvp_native_vs_jac_corrected, wcvp_native_vs_bcd_corrected,wcvp_all_vs_pi_corrected, wcvp_all_vs_jac_corrected, wcvp_all_vs_bcd_corrected, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("wcvp_popsize_vs_diversity_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

### supplemental figures
ggarrange(wcvp_native_area_vs_pi_corrected, wcvp_native_area_vs_jac_corrected, wcvp_native_area_vs_bcd_corrected,wcvp_all_area_vs_pi_corrected, wcvp_all_area_vs_jac_corrected, wcvp_all_area_vs_bcd_corrected, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("wcvp_area_vs_diversity_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(gbif_native_vs_pi_corrected, gbif_native_vs_jac_corrected, gbif_native_vs_bcd_corrected,gbif_all_vs_pi_corrected, gbif_all_vs_jac_corrected, gbif_all_vs_bcd_corrected, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_popsize_vs_diversity_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(gbif_native_area_vs_pi_corrected, gbif_native_area_vs_jac_corrected, gbif_native_area_vs_bcd_corrected,gbif_all_area_vs_pi_corrected, gbif_all_area_vs_jac_corrected, gbif_all_area_vs_bcd_corrected, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_area_vs_diversity_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(height_vs_pi_corrected, height_vs_jac_corrected, height_vs_bcd_corrected, nrow = 1, ncol = 3, labels = c("A", "B", "C"), font.label = list(size = 12))

ggsave(paste("height_vs_diversity_corrected_", today, ".jpg", sep = ""), height = 3, width = 7.5, units = "in", dpi = 350)

## population size vs diversity, including genome size

### Figure 3
ggarrange(pop_genome_wcvp_all_vs_pi_figure, pop_genome_wcvp_all_vs_bcd_figure, nrow = 1, ncol = 2, labels = c("A", "B"), font.label = list(size = 12))

ggsave(paste("wcvp_all_popsize_vs_pi_and_bcd_genome-size-corrected_", today, ".tiff", sep = ""), height = 3, width = 5, units = "in", compression = "lzw")
ggsave(paste("wcvp_all_popsize_vs_pi_and_bcd_genome-size-corrected_", today, ".jpg", sep = ""), height = 3, width = 5, units = "in", dpi = 350)

ggarrange(pop_genome_wcvp_all_vs_jac_figure)

ggsave(paste("wcvp_all_popsize_vs_jac_genome-size-corrected_", today, ".jpg", sep = ""), height = 3, width = 3, units = "in", dpi = 350)

ggarrange(pop_genome_wcvp_native_vs_pi_figure, pop_genome_wcvp_native_vs_jac_figure, pop_genome_wcvp_native_vs_bcd_figure, nrow = 1, ncol = 3, labels = c("A", "B", "C"), font.label = list(size = 12))

ggsave(paste("wcvp_native_popsize_vs_diversity_genome-size-corrected_", today, ".jpg", sep = ""), height = 3, width = 7.5, units = "in", dpi = 350)

### supplemental figures
ggarrange(pop_genome_gbif_native_vs_pi_figure, pop_genome_gbif_native_vs_jac_figure, pop_genome_gbif_native_vs_bcd_figure,pop_genome_gbif_all_vs_pi_figure, pop_genome_gbif_all_vs_jac_figure, pop_genome_gbif_all_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_popsize_vs_diversity_genome-size-corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(pop_genome_wcvp_native_area_vs_pi_figure, pop_genome_wcvp_native_area_vs_jac_figure, pop_genome_wcvp_native_area_vs_bcd_figure,pop_genome_wcvp_all_area_vs_pi_figure, pop_genome_wcvp_all_area_vs_jac_figure, pop_genome_wcvp_all_area_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("wcvp_area_vs_diversity_genome-size-corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(pop_genome_gbif_native_area_vs_pi_figure, pop_genome_gbif_native_area_vs_jac_figure, pop_genome_gbif_native_area_vs_bcd_figure,pop_genome_gbif_all_area_vs_pi_figure, pop_genome_gbif_all_area_vs_jac_figure, pop_genome_gbif_all_area_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_area_vs_diversity_genome-size-corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(pop_genome_height_vs_pi_figure, pop_genome_height_vs_jac_figure, pop_genome_height_vs_bcd_figure, nrow = 1, ncol = 3, labels = c("A", "B", "C"), font.label = list(size = 12))

ggsave(paste("height_vs_diversity_genome-size-corrected_", today, ".jpg", sep = ""), height = 4, width = 7.5, units = "in", dpi = 350)

## genome size vs diversity

### Figure 4
ggarrange(genome_wcvp_all_vs_pi_figure, genome_wcvp_all_vs_bcd_figure, nrow = 1, ncol = 2, labels = c("A", "B"), font.label = list(size = 12))

ggsave(paste("pi_bcd_vs_genome_size_wcvp-all-corrected_", today, ".tiff", sep = ""), height = 3, width = 5, units = "in", compression = "lzw")
ggsave(paste("pi_bcd_vs_genome_size_wcvp-all-corrected_", today, ".jpg", sep = ""), height = 3, width = 5, units = "in", dpi = 350)

ggarrange(genome_wcvp_all_vs_jac_figure)

ggsave(paste("jac_vs_genome_size_wcvp-all-corrected_", today, ".jpg", sep = ""), height = 3, width = 3, units = "in", dpi = 350)

ggarrange(genome_wcvp_native_vs_pi_figure, genome_wcvp_native_vs_jac_figure, genome_wcvp_native_vs_bcd_figure, nrow = 1, ncol = 3, labels = c("A", "B", "C"), font.label = list(size = 12))

ggsave(paste("pi_jac_bcd_vs_genome_size_wcvp-native-corrected_", today, ".jpg", sep = ""), height = 3, width = 7.5, units = "in", dpi = 350)

### supplemental figures

ggarrange(genome_gbif_native_vs_pi_figure, genome_gbif_native_vs_jac_figure, genome_gbif_native_vs_bcd_figure,genome_gbif_all_vs_pi_figure, genome_gbif_all_vs_jac_figure, genome_gbif_all_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_popsize_vs_genome_size_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(genome_wcvp_native_area_vs_pi_figure, genome_wcvp_native_area_vs_jac_figure, genome_wcvp_native_area_vs_bcd_figure,genome_wcvp_all_area_vs_pi_figure, genome_wcvp_all_area_vs_jac_figure, genome_wcvp_all_area_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("wcvp_area_vs_genome_size_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(genome_gbif_native_area_vs_pi_figure, genome_gbif_native_area_vs_jac_figure, genome_gbif_native_area_vs_bcd_figure,genome_gbif_all_area_vs_pi_figure, genome_gbif_all_area_vs_jac_figure, genome_gbif_all_area_vs_bcd_figure, nrow = 2, ncol = 3, labels = c("A", "B", "C", "D", "E", "F"), font.label = list(size = 12))

ggsave(paste("gbif_area_vs_genome_size_corrected_", today, ".jpg", sep = ""), height = 6, width = 7.5, units = "in", dpi = 350)

ggarrange(genome_height_vs_pi_figure, genome_height_vs_jac_figure, genome_height_vs_bcd_figure, nrow = 1, ncol = 3, labels = c("A", "B", "C"), font.label = list(size = 12))

ggsave(paste("height_vs_genome_size_corrected_", today, ".jpg", sep = ""), height = 4, width = 7.5, units = "in", dpi = 350)

#### save results of phylogenetic least squares models as tables ####

mod_coeff_results = rbind(
  summary(mod_gbif_all_popsize_pi)$coefficients,
  summary(mod_gbif_all_popsize_jac)$coefficients,
  summary(mod_gbif_all_popsize_bcd)$coefficients,
  summary(mod_gbif_all_area_pi)$coefficients,
  summary(mod_gbif_all_area_jac)$coefficients,
  summary(mod_gbif_all_area_bcd)$coefficients,
  summary(mod_gbif_native_popsize_pi)$coefficients,
  summary(mod_gbif_native_popsize_jac)$coefficients,
  summary(mod_gbif_native_popsize_bcd)$coefficients,
  summary(mod_gbif_native_area_pi)$coefficients,
  summary(mod_gbif_native_area_jac)$coefficients,
  summary(mod_gbif_native_area_bcd)$coefficients,
  summary(mod_wcvp_all_popsize_pi)$coefficients,
  summary(mod_wcvp_all_popsize_jac)$coefficients,
  summary(mod_wcvp_all_popsize_bcd)$coefficients,
  summary(mod_wcvp_all_area_pi)$coefficients,
  summary(mod_wcvp_all_area_jac)$coefficients,
  summary(mod_wcvp_all_area_bcd)$coefficients,
  summary(mod_wcvp_native_popsize_pi)$coefficients,
  summary(mod_wcvp_native_popsize_jac)$coefficients,
  summary(mod_wcvp_native_popsize_bcd)$coefficients,
  summary(mod_wcvp_native_area_pi)$coefficients,
  summary(mod_wcvp_native_area_jac)$coefficients,
  summary(mod_wcvp_native_area_bcd)$coefficients,
  summary(mod_height_pi)$coefficients,
  summary(mod_height_jac)$coefficients,
  summary(mod_height_bcd)$coefficients,
  summary(genome_wcvp_all_vs_pi)$coefficients,
  summary(genome_wcvp_all_vs_jac)$coefficients,
  summary(genome_wcvp_all_vs_bcd)$coefficients,
  summary(genome_wcvp_native_vs_pi)$coefficients,
  summary(genome_wcvp_native_vs_jac)$coefficients,
  summary(genome_wcvp_native_vs_bcd)$coefficients,
  summary(genome_wcvp_all_area_vs_pi)$coefficients,
  summary(genome_wcvp_all_area_vs_jac)$coefficients,
  summary(genome_wcvp_all_area_vs_bcd)$coefficients,
  summary(genome_wcvp_native_area_vs_pi)$coefficients,
  summary(genome_wcvp_native_area_vs_jac)$coefficients,
  summary(genome_wcvp_native_area_vs_bcd)$coefficients,
  summary(genome_gbif_all_vs_pi)$coefficients,
  summary(genome_gbif_all_vs_jac)$coefficients,
  summary(genome_gbif_all_vs_bcd)$coefficients,
  summary(genome_gbif_native_vs_pi)$coefficients,
  summary(genome_gbif_native_vs_jac)$coefficients,
  summary(genome_gbif_native_vs_bcd)$coefficients,
  summary(genome_gbif_all_area_vs_pi)$coefficients,
  summary(genome_gbif_all_area_vs_jac)$coefficients,
  summary(genome_gbif_all_area_vs_bcd)$coefficients,
  summary(genome_gbif_native_area_vs_pi)$coefficients,
  summary(genome_gbif_native_area_vs_jac)$coefficients,
  summary(genome_gbif_native_area_vs_bcd)$coefficients,
  summary(genome_height_vs_pi)$coefficients,
  summary(genome_height_vs_jac)$coefficients,
  summary(genome_height_vs_bcd)$coefficients
)

# create column from row names
mod_coeff_results = cbind(mod_coeff_results, rownames(mod_coeff_results))

# convert to dataframe
mod_coeff_results = as.data.frame(mod_coeff_results)

# add some extra descriptors: name of response variable and model id, so that readers know which values are all together in one model
mod_coeff_results$Response = c(
  rep(mod_gbif_all_popsize_pi$namey, times = 5),
  rep(mod_gbif_all_popsize_jac$namey, times = 5),
  rep(mod_gbif_all_popsize_bcd$namey, times = 5),
  rep(mod_gbif_all_area_pi$namey, times = 5),
  rep(mod_gbif_all_area_jac$namey, times = 5),
  rep(mod_gbif_all_area_bcd$namey, times = 5),
  rep(mod_gbif_native_popsize_pi$namey, times = 5),
  rep(mod_gbif_native_popsize_jac$namey, times = 5),
  rep(mod_gbif_native_popsize_bcd$namey, times = 5),
  rep(mod_gbif_native_area_pi$namey, times = 5),
  rep(mod_gbif_native_area_jac$namey, times = 5),
  rep(mod_gbif_native_area_bcd$namey, times = 5),
  rep(mod_wcvp_all_popsize_pi$namey, times = 5),
  rep(mod_wcvp_all_popsize_jac$namey, times = 5),
  rep(mod_wcvp_all_popsize_bcd$namey, times = 5),
  rep(mod_wcvp_all_area_pi$namey, times = 5),
  rep(mod_wcvp_all_area_jac$namey, times = 5),
  rep(mod_wcvp_all_area_bcd$namey, times = 5),
  rep(mod_wcvp_native_popsize_pi$namey, times = 5),
  rep(mod_wcvp_native_popsize_jac$namey, times = 5),
  rep(mod_wcvp_native_popsize_bcd$namey, times = 5),
  rep(mod_wcvp_native_area_pi$namey, times = 5),
  rep(mod_wcvp_native_area_jac$namey, times = 5),
  rep(mod_wcvp_native_area_bcd$namey, times = 5),
  rep(mod_height_pi$namey, times = 5),
  rep(mod_height_jac$namey, times = 5),
  rep(mod_height_bcd$namey, times = 5),
  rep(genome_wcvp_all_vs_pi$namey, times = 6),
  rep(genome_wcvp_all_vs_jac$namey, times = 6),
  rep(genome_wcvp_all_vs_bcd$namey, times = 6),
  rep(genome_wcvp_native_vs_pi$namey, times = 6),
  rep(genome_wcvp_native_vs_jac$namey, times = 6),
  rep(genome_wcvp_native_vs_bcd$namey, times = 6),
  rep(genome_wcvp_all_area_vs_pi$namey, times = 6),
  rep(genome_wcvp_all_area_vs_jac$namey, times = 6),
  rep(genome_wcvp_all_area_vs_bcd$namey, times = 6),
  rep(genome_wcvp_native_area_vs_pi$namey, times = 6),
  rep(genome_wcvp_native_area_vs_jac$namey, times = 6),
  rep(genome_wcvp_native_area_vs_bcd$namey, times = 6),
  rep(genome_gbif_all_vs_pi$namey, times = 6),
  rep(genome_gbif_all_vs_jac$namey, times = 6),
  rep(genome_gbif_all_vs_bcd$namey, times = 6),
  rep(genome_gbif_native_vs_pi$namey, times = 6),
  rep(genome_gbif_native_vs_jac$namey, times = 6),
  rep(genome_gbif_native_vs_bcd$namey, times = 6),
  rep(genome_gbif_all_area_vs_pi$namey, times = 6),
  rep(genome_gbif_all_area_vs_jac$namey, times = 6),
  rep(genome_gbif_all_area_vs_bcd$namey, times = 6),
  rep(genome_gbif_native_area_vs_pi$namey, times = 6),
  rep(genome_gbif_native_area_vs_jac$namey, times = 6),
  rep(genome_gbif_native_area_vs_bcd$namey, times = 6),
  rep(genome_height_vs_pi$namey, times = 6),
  rep(genome_height_vs_jac$namey, times = 6),
  rep(genome_height_vs_bcd$namey, times = 6)
)

mod_coeff_results$ID = c(
  rep(1, times = 5),
  rep(2, times = 5),
  rep(3, times = 5),
  rep(4, times = 5),
  rep(5, times = 5),
  rep(6, times = 5),
  rep(7, times = 5),
  rep(8, times = 5),
  rep(9, times = 5),
  rep(10, times = 5),
  rep(11, times = 5),
  rep(12, times = 5),
  rep(13, times = 5),
  rep(14, times = 5),
  rep(15, times = 5),
  rep(16, times = 5),
  rep(17, times = 5),
  rep(18, times = 5),
  rep(19, times = 5),
  rep(20, times = 5),
  rep(21, times = 5),
  rep(22, times = 5),
  rep(23, times = 5),
  rep(24, times = 5),
  rep(25, times = 5),
  rep(26, times = 5),
  rep(27, times = 5),
  rep(28, times = 6),
  rep(29, times = 6),
  rep(30, times = 6),
  rep(31, times = 6),
  rep(32, times = 6),
  rep(33, times = 6),
  rep(34, times = 6),
  rep(35, times = 6),
  rep(36, times = 6),
  rep(37, times = 6),
  rep(38, times = 6),
  rep(39, times = 6),
  rep(40, times = 6),
  rep(41, times = 6),
  rep(42, times = 6),
  rep(43, times = 6),
  rep(44, times = 6),
  rep(45, times = 6),
  rep(46, times = 6),
  rep(47, times = 6),
  rep(48, times = 6),
  rep(49, times = 6),
  rep(50, times = 6),
  rep(51, times = 6),
  rep(52, times = 6),
  rep(53, times = 6),
  rep(54, times = 6)
)

#mod_coeff_results$Response = rep(
#    c(
#      rep("4-fold SNP diversity", times = 5),
#      rep("Jaccard dissimilarity", times = 5),
#      rep("Bray-Curtis dissimilarity", times = 5)
#    ),
#    times = 9
#)

#mod_coeff_results$`Population size metric` = c(
#  rep("GBIF", times = 30),
#  rep("WCVP", times = 30)
#)

#mod_coeff_results$`Range included` = c(
#  rep("full", times = 15),
#  rep("native only", times = 15),
#  rep("full", times = 15),
#  rep("native only", times = 15)
#)

# reorganize the columns
mod_coeff_results = mod_coeff_results[,c(7,6,5,1:4)]

colnames(mod_coeff_results)[3] = "Variable"

# save results
write.csv(mod_coeff_results, paste("model_coefficients_", today, ".csv", sep = ""), row.names = F)

#### Save whole model statistics as a separate table ####

# Write function to extract statistics that apply to entire model
extract_mod_results = function(mod){
  mod_sum = summary(mod)
  mod_frame = data.frame(
    call = as.character(mod_sum$call)[2],
    fstatvalue = mod_sum$fstatistic["value"],
    fstatnumdf = mod_sum$fstatistic["numdf"],
    fstatdendf = mod_sum$fstatistic["dendf"],
    rsq = mod_sum$r.squared,
    adjrsq = mod_sum$adj.r.squared[1,1]
  )
  return(mod_frame)
}

whole_mod_results = rbind(
  extract_mod_results(mod_gbif_all_popsize_pi),
  extract_mod_results(mod_gbif_all_popsize_jac),
  extract_mod_results(mod_gbif_all_popsize_bcd),
  extract_mod_results(mod_gbif_all_area_pi),
  extract_mod_results(mod_gbif_all_area_jac),
  extract_mod_results(mod_gbif_all_area_bcd),
  extract_mod_results(mod_gbif_native_popsize_pi),
  extract_mod_results(mod_gbif_native_popsize_jac),
  extract_mod_results(mod_gbif_native_popsize_bcd),
  extract_mod_results(mod_gbif_native_area_pi),
  extract_mod_results(mod_gbif_native_area_jac),
  extract_mod_results(mod_gbif_native_area_bcd),
  extract_mod_results(mod_wcvp_all_popsize_pi),
  extract_mod_results(mod_wcvp_all_popsize_jac),
  extract_mod_results(mod_wcvp_all_popsize_bcd),
  extract_mod_results(mod_wcvp_all_area_pi),
  extract_mod_results(mod_wcvp_all_area_jac),
  extract_mod_results(mod_wcvp_all_area_bcd),
  extract_mod_results(mod_wcvp_native_popsize_pi),
  extract_mod_results(mod_wcvp_native_popsize_jac),
  extract_mod_results(mod_wcvp_native_popsize_bcd),
  extract_mod_results(mod_wcvp_native_area_pi),
  extract_mod_results(mod_wcvp_native_area_jac),
  extract_mod_results(mod_wcvp_native_area_bcd),
  extract_mod_results(mod_height_pi),
  extract_mod_results(mod_height_jac),
  extract_mod_results(mod_height_bcd),
  extract_mod_results(genome_wcvp_all_vs_pi),
  extract_mod_results(genome_wcvp_all_vs_jac),
  extract_mod_results(genome_wcvp_all_vs_bcd),
  extract_mod_results(genome_wcvp_native_vs_pi),
  extract_mod_results(genome_wcvp_native_vs_jac),
  extract_mod_results(genome_wcvp_native_vs_bcd),
  extract_mod_results(genome_wcvp_all_area_vs_pi),
  extract_mod_results(genome_wcvp_all_area_vs_jac),
  extract_mod_results(genome_wcvp_all_area_vs_bcd),
  extract_mod_results(genome_wcvp_native_area_vs_pi),
  extract_mod_results(genome_wcvp_native_area_vs_jac),
  extract_mod_results(genome_wcvp_native_area_vs_bcd),
  extract_mod_results(genome_gbif_all_vs_pi),
  extract_mod_results(genome_gbif_all_vs_jac),
  extract_mod_results(genome_gbif_all_vs_bcd),
  extract_mod_results(genome_gbif_native_vs_pi),
  extract_mod_results(genome_gbif_native_vs_jac),
  extract_mod_results(genome_gbif_native_vs_bcd),
  extract_mod_results(genome_gbif_all_area_vs_pi),
  extract_mod_results(genome_gbif_all_area_vs_jac),
  extract_mod_results(genome_gbif_all_area_vs_bcd),
  extract_mod_results(genome_gbif_native_area_vs_pi),
  extract_mod_results(genome_gbif_native_area_vs_jac),
  extract_mod_results(genome_gbif_native_area_vs_bcd),
  extract_mod_results(genome_height_vs_pi),
  extract_mod_results(genome_height_vs_jac),
  extract_mod_results(genome_height_vs_bcd)
)

whole_mod_results$ID = 1:54

names(whole_mod_results) = c("Model", "F statistic", "F test df1", "F test df2", "R squared", "Adjusted R squared", "ID")

whole_mod_results = whole_mod_results[,c(7,1:6)]

write.csv(whole_mod_results, paste("model_rsquared_", today, ".csv", sep = ""), row.names = F, quote = F)

```

# plot phylogeny alongside data

ggtree causes linear models to crash, so need to do this separately

```{r}

#if (!require("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")
#BiocManager::install("ggtree")

library(ggtree)

# test ggtree package
#tr = rtree(10)
#ggtree(tr) + geom_tiplab()

# arrange plot data to match order of tip labels
plotdf = alldiv %>% arrange(factor(species, levels = phy$tip.label))

plotdf$Ploidy = as.numeric(plotdf$Ploidy)

plotdf = plotdf[!is.na(plotdf$species),]

plotdf$retained = (plotdf$nvariant >= 1000 & plotdf$meancov > 0.5)

# remove species from phylogeny that lack data
plotphy = drop.tip(phy, tip = phy$tip.label[!(phy$tip.label %in% alldiv$species)])

# change tip labels to be mating system
#plotphy$tip.label = plotdf$Mating.system

 # get color pallete
#plotpal = scico(5, palette = "lipari")
plotpal = c("#021326", "#3B5378", "#7F5F70", "#CE685E", "#E5AA7F")
#plotpal = c("#021326","#294B70","#6B5F76","#A36267","#E57A61","#E5B488","#FDF4D9")
#plotpal = c("#021326","#3B5378","#7F5F70","#CE685E","#E5AA7F","#FDF4D9")
#plotpal = c("#191900","#663329","#D85F4D","#ECAC54","#FFFECB")

# vector of colors to show outlier vs not
# scico(2, palette = "bam")
outlier_colors = plotdf$retained
outlier_colors[outlier_colors] = "#0C4C00"
outlier_colors[(outlier_colors == FALSE)] = "#65014B"

# plot phylogeny alongside other data
ggtree(plotphy, size = 1.5) + 
  geom_tiplab(color = outlier_colors, size = 2, offset = 10) +
  #geom_tippoint(aes(shape = retained, node = NULL), data = plotdf, size = 1.5) +
  #geom_facet(panel = "Retained", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = x, color = retained)) +
  geom_facet(panel = "log(WCVP full ratio)", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = log10(wcvp_all_popsize)), color = plotpal[2]) +
  geom_facet(panel = "log(4-fold SNP diversity)", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = log10(pi)), color = plotpal[3]) +
  geom_facet(panel = "Bray-Curtis dissimilarity", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = bcd), color = plotpal[4]) +
  #geom_facet(panel = "Jaccard dissimilarity", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = jac), color = plotpal[5]) +
  theme_tree2() +
  theme(strip.background = element_rect(fill="white")) +
  theme(strip.text = element_text(colour = 'black'), text = element_text(size = 12)) +
  xlim_tree(550)
  
ggsave(paste("tree_with_primary_variables_", today, ".tiff", sep = ""), width = 7.5, height = 7.75, compression = "lzw")
ggsave(paste("tree_with_primary_variables_", today, ".jpg", sep = ""), width = 7.5, height = 7.75, dpi = 350)

# plot technical variables
ggtree(plotphy, size = 1.5) +
  geom_tiplab(size = 2) +
  #geom_tippoint(mapping = aes(shape = Mating.system), data = plotdf, size = 1.5) + 
  #geom_facet(panel = "Ploidy", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = Ploidy), color = plotpal[1]) +
  #geom_facet(panel = "Genome size", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = genome.size), color = plotpal[2]) +
  geom_facet(panel = "Individuals", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = nidv), color = plotpal[3]) +
  geom_facet(panel = "Mean coverage", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = (totalbp/(genome.size*nidv))), color = plotpal[4]) +
  geom_facet(panel = "CV in bp sequenced", data = plotdf, geom = geom_segment, size = 3, mapping = aes(x = 0, y = y, yend = y, xend = cvbp), color = plotpal[5]) +
  theme_tree2() +
  theme(strip.background = element_rect(fill="white")) +
  theme(strip.text = element_text(colour = 'black'), text = element_text(size = 12)) +
  xlim_tree(500)

ggsave(paste("tree_with_technical_variables_", today, ".jpg", sep = ""), width = 7.5, height = 7.75, dpi = 350)

```
